Implement HardwareSerial class begin, write, read and available mathods. Need to implement usart ISR and global buffer

+++ include/classes/RemoteXY_Serial.h
@@ -11,20 +11,20 @@ public:
 #if defined(REMOTEXY_PORT__HARDSERIAL)
 
 	HardwareSerial * serial;
-	void initSerial (HardwareSerial * _serial, long _serialSpeed) {
+	void initSerial (HardwareSerial * _serial, uint32_t _serialSpeed) {
 		serial = _serial;
 		serial->begin (_serialSpeed);
 	}
 #elif defined(REMOTEXY_PORT__SOFTSERIAL)
 #if defined(SoftwareSerial_h)
 	SoftwareSerial * serial;
-	void initSerial (uint8_t _serialRx, uint8_t _serialTx, long _serialSpeed) {
+	void initSerial (uint8_t _serialRx, uint8_t _serialTx, uint32_t _serialSpeed) {
 		serial = new SoftwareSerial (_serialRx, _serialTx);
 		serial->begin (_serialSpeed);
 	}
 #elif defined(SoftSerial_h)
 	SoftSerial * serial;
-	void initSerial (uint8_t _serialRx, uint8_t _serialTx, long _serialSpeed) {
+	void initSerial (uint8_t _serialRx, uint8_t _serialTx, uint32_t _serialSpeed) {
 		serial = new SoftSerial (_serialRx, _serialTx);
 		serial->begin (_serialSpeed);
 	}
